// libs/juridico/src/lib/models/SeccionDeterminativa/detDeterminacionDeudaModel.ts
export interface DetDeterminacionDeudaModel {
  refTabla: string;
  numFormul: string;
  mtoCas137: number;
  mtoCas138: number;
  mtoCas139: number;
  mtoCas142: number;
  mtoCas505: number;
  mtoCas141: number;
  mtoCas144: number;
  mtoCas145: number;
  mtoCas146: number;
  mtoCas180: number;
}

export interface DeudaPagosPreviosModel {
  codTri: string,
  numOrd: string,
  fecPres: string,
  mtoPag: number,
  indSel: string,
  codFor: string
}