// libs/juridico/src/lib/models/SeccionDeterminativa/detEstFinancierosModel.ts
export interface ActivoEmp {
    refTabla: string;
    numFormul: string;
    mtoCas359: number;
    mtoCas360: number;
    mtoCas361: number;
    mtoCas362: number;
    mtoCas363: number;
    mtoCas364: number;
    mtoCas365: number;
    mtoCas366: number;
    mtoCas367: number;
    mtoCas368: number;
    mtoCas369: number;
    mtoCas370: number;
    mtoCas371: number;
    mtoCas372: number;
    mtoCas373: number;
    mtoCas374: number;
    mtoCas375: number;
    mtoCas376: number;
    mtoCas377: number;
    mtoCas378: number;
    mtoCas379: number;
    mtoCas380: number;
    mtoCas381: number;
    mtoCas382: number;
    mtoCas383: number;
    mtoCas384: number;
    mtoCas385: number;
    mtoCas386: number;
    mtoCas387: number;
    mtoCas388: number;
    mtoCas389: number;
    mtoCas390: number;

}
export interface PasivoPatrEmp {
    refTabla: string;
    numFormul: string;
    mtoCas401: number;
    mtoCas402: number;
    mtoCas403: number;
    mtoCas404: number;
    mtoCas405: number;
    mtoCas406: number;
    mtoCas407: number;
    mtoCas408: number;
    mtoCas409: number;
    mtoCas410: number;
    mtoCas411: number;
    mtoCas412: number;
    mtoCas414: number;
    mtoCas415: number;
    mtoCas416: number;
    mtoCas417: number;
    mtoCas418: number;
    mtoCas419: number;
    mtoCas420: number;
    mtoCas421: number;
    mtoCas422: number;
    mtoCas423: number;
    mtoCas424: number;
    mtoCas425: number;
    mtoCas426: number;

}
export interface Ganancia {
    refTabla: string;
    numFormul: string;
    mtoCas461: number;
    mtoCas462: number;
    mtoCas463: number;
    mtoCas464: number;
    mtoCas466: number;
    mtoCas467: number;
    mtoCas468: number;
    mtoCas469: number;
    mtoCas470: number;
    mtoCas471: number;
    mtoCas472: number;
    mtoCas473: number;
    mtoCas475: number;
    mtoCas476: number;
    mtoCas477: number;
    mtoCas478: number;
    mtoCas480: number;
    mtoCas481: number;
    mtoCas483: number;
    mtoCas484: number;
    mtoCas485: number;
    mtoCas486: number;
    mtoCas487: number;
    mtoCas489: number;
    mtoCas490: number;
    mtoCas492: number;
    mtoCas493: number;
    mtoCas650: number;
    mtoCas651: number;
    mtoCas432: number;
    mtoCas433: number;
}